+++ lib/droidknightsapp_home.dart
@@ -1,6 +1,9 @@
 import 'package:droidknights/pages/track_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:droidknights/pages/info_page.dart';
+import 'package:flutter/cupertino.dart';
+
+import 'dart:io' show Platform;
 
 class DroidknightsAppHome extends StatefulWidget {
   @override
@@ -11,27 +14,22 @@ class _DroidknightsAppHomeState extends State<DroidknightsAppHome>
     with TickerProviderStateMixin {
   TabController _tabController;
   int _currentIndex = 0;
-  List<Widget> _appbar = [];
-  List<Widget> _children = [];
+  List<Widget> _children = [
+    InfoPage(),
+    TrackScreen(),
+  ];
 
   @override
   void initState() {
     super.initState();
     _tabController = new TabController(length: 4, vsync: this, initialIndex: 0);
-    _appbar.addAll([null, scheduleAppbar()]);
-    _children.addAll([
-      InfoPage(),
-      //PlaceholderWidget(Colors.deepOrange),
-      scheduleBody()
-    ]);
   }
 
   @override
   Widget build(BuildContext context) {
-    return new Scaffold(
-        appBar: appBar(),
-        bottomNavigationBar: bottomNavigationBar(),
-        body: _children[_currentIndex]);
+    return Platform.isAndroid
+        ? createAndroidWidget(context)
+    : createIosWidget(context);
   }
 
   void onTabTapped(int index) {
@@ -40,56 +38,62 @@ class _DroidknightsAppHomeState extends State<DroidknightsAppHome>
     });
   }
 
-  Widget appBar() => new AppBar(
-    title: Row(
-    mainAxisAlignment: MainAxisAlignment.center,
-    children: [
-      Image.asset(
-        'assets/images/dk_appbar_title.png',
-        fit: BoxFit.contain,
-        height: 25,
-      )
-    ],),
-    elevation: 0.7,
-    bottom: _appbar[_currentIndex]
-  );
-
-  Widget bottomNavigationBar() => new BottomNavigationBar(
-    onTap: onTabTapped,
-    currentIndex: _currentIndex,
-    // this will be set when a new tab is tapped
-    items: [
-      BottomNavigationBarItem(
-        icon: new Icon(Icons.info),
-        title: new Text('Info'),
-      ),
-      BottomNavigationBarItem(
-        icon: new Icon(Icons.schedule),
-        title: new Text('Schedule'),
-      ),
-    ]
-  );
-
-  Widget scheduleAppbar() {
-    return new TabBar(
-        controller: _tabController,
-        labelColor: new Color(0xff40d225),
-        unselectedLabelColor: Colors.grey,
-        indicatorColor: new Color(0xff40d225),
-        tabs: <Widget>[
-          new Tab(text: "Track1"),
-          new Tab(text: "Track2"),
-          new Tab(text: "Track3"),
-        ]);
+  Widget createAndroidWidget(BuildContext context) {
+    return new Scaffold(
+        bottomNavigationBar: bottomNavigationBar(),
+        body: _children[_currentIndex]);
   }
 
-  Widget scheduleBody() {
-    return new Scaffold(
-      body: new TabBarView(controller: _tabController, children: <Widget>[
-        new TrackScreen('assets/json/schedule_track1.json'),
-        new TrackScreen('assets/json/schedule_track2.json'),
-        new TrackScreen('assets/json/schedule_track3.json')
-      ]),
+  Widget createIosWidget(BuildContext context) {
+    return CupertinoTabScaffold(
+      backgroundColor:  const Color(0xFF112030),
+      tabBar: iosNavigationBar(),
+      tabBuilder: (context, index) {
+        return CupertinoTabView(
+          builder: (context) {
+            switch (index) {
+              case 0:
+                return _children[0];
+                break;
+              case 1:
+                return _children[1];
+                break;
+              default:
+                return Center(child: Text("DEFAULT"));
+            }
+          },
+        );
+      },
     );
   }
+
+  Widget iosNavigationBar() => CupertinoTabBar(
+        backgroundColor: CupertinoColors.lightBackgroundGray,
+        items: [
+          BottomNavigationBarItem(
+            icon: Icon(CupertinoIcons.info),
+            title: Text('Info'),
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(CupertinoIcons.clock),
+            title: Text('Schedule'),
+          ),
+        ],
+      );
+
+
+
+  Widget bottomNavigationBar() =>
+      new BottomNavigationBar(onTap: onTabTapped, currentIndex: _currentIndex,
+          // this will be set when a new tab is tapped
+          items: [
+            BottomNavigationBarItem(
+              icon: new Icon(Icons.info),
+              title: new Text('Info'),
+            ),
+            BottomNavigationBarItem(
+              icon: new Icon(Icons.schedule),
+              title: new Text('Schedule'),
+            ),
+          ]);
 }
